Resume
This is a simple javaFX project. It is a Game which main character is a snake. The game contains 5 levels. The snake can be controlled 
with W (up), S(down), R(right), L(left). The game can be "Paused/Unpaused" everytime with button "P" and it can be "Restarted" with button "R" only if there is a "Game Over" label on the screen. Snake's speed increases sometimes. There is a "magic" (blue) box which appears on a period of time. Its power is to decrease snake's speed for a short time. On the last level there is no limit about playing - the only problem which can cause "Game Over" is the snake's self-destroy. Be careful with the things in RED color. If the snake touches one of these things, the label "Game Over" will appear on the screen.